Optimisation (#235)
Moved SceneRenderNodes and SceneCulledRenderNodes to their own files.  
Moved UBOs to UBO pools.  
Speed improvement for directional lights shadow maps.  
Fixed a bit shadow maps update.  
Removed screen buffer management in Win32 console.  
Split update and render for shadow maps.  
Disabled fence in light passes.  
Moved volumetricDither to a static constant c3d_volumetricDither.  
Removed remaining cuT() in shaders.  
Replaced MatrixUbo with ShadowConfig, for directional light.  
Fixed CSM.  
Debug rainbooooow !  
Improved binary loader logging.  
Added tinyexr to load EXR image format.  
Fixed export crash  
Fixed a bit mesh converter.  
Fixed some bugs when mesh loading fails.  
Fixed type for normal_factor and height_factor parsers.  
Fixed filter in RenderQuad.  
Fixed SMAA with OpenGL renderer.  
Added image format lookup helper function, for stencil formats.  
Fixed SMAA with Direct3D.  
Fixed inconsistency with diffuse and ambient between deferred and forward.

+++ source/Plugins/PostEffects/SmaaPostEffect/BlendingWeightCalculation.hpp
@@ -11,7 +11,7 @@ See LICENSE file in root folder
 #include <Castor3D/Render/PostEffect/PostEffectSurface.hpp>
 #include <Castor3D/Render/ToTexture/RenderQuad.hpp>
 
-#include <ShaderWriter/Shader.hpp>
+#include <ShaderAST/Shader.hpp>
 
 namespace smaa
 {
@@ -25,7 +25,7 @@ namespace smaa
 			, SmaaConfig const & config );
 		castor3d::CommandsSemaphore prepareCommands( castor3d::RenderPassTimer const & timer
 			, uint32_t passIndex );
-		void update( castor::Point4i const & subsampleIndices );
+		void update( castor::Point4f const & subsampleIndices );
 		void accept( castor3d::PipelineVisitorBase & visitor );
 
 		inline castor3d::TextureLayoutSPtr getSurface()const
@@ -42,7 +42,7 @@ namespace smaa
 		ashes::RenderPassPtr m_renderPass;
 		castor3d::PostEffectSurface m_surface;
 		ashes::SamplerPtr m_pointSampler;
-		castor3d::UniformBufferUPtr< castor::Point4i > m_ubo;
+		castor3d::UniformBufferUPtr< castor::Point4f > m_ubo;
 		castor3d::TextureLayoutSPtr m_areaTex;
 		castor3d::TextureLayoutSPtr m_searchTex;
 		castor3d::ShaderModule m_vertexShader;
